2.1. Setting up Webpack Encore--composer require encore--yarn install
So let's get Webpack Encore installed so we can get a proper CSS and JavaScript build system set up.

Installing Encore
Find your terminal and run:

composer require encore
This really installs WebpackEncoreBundle, which will give us a few Twig helper functions. But its true superpower is its recipe.

The Encore Recipe Files
After Composer finishes, run:

git status
Ok: it modified several files, most of which make sense. The .gitignore file is now ignoring the node_modules/ directory and the bundle was automatically enabled in config/bundles.php.

There are also several new config files, which mostly aren't too important. What is important is that it created a package.json file. Let's go check that out. This defines several things that our Webpack Encore setup needs, like Encore itself and stimulus, which we'll talk more about later.

The other hugely important file the recipe added was webpack.config.js - with a basic setup. The recipe also added this assets/ directory. There are a few things in here related to Stimulus: the controllers stuff and bootstrap.js. We'll talk about those soon.

To download the dependencies described in package.json, find your terminal and run:

yarn install
Oh, and if you're totally new to Encore and want to dig into it deeper, check out our Webpack Encore: Frontend like a Pro! course.

The Entry Files and First Build
Right now, our app has a pretty standard Webpack Encore setup. We have one entry which is this assets/app.js file. That means that when we build our assets in a few minutes, Webpack will only look at this file to figure out all the JavaScript and CSS needed in the project. It will then output the final, built CSS and JavaScript files into a public/build/ directory.

Of course, when it reads this file, it will follow any imports, like this import of ./styles/app.css... which is a dummy file that sets the background-color to gray.

So, once we have executed Encore and built the assets - which we'll do in a minute - we're going to need script and link tags that point to those new files.