2.2. Setting up Webpack Encore
Let's add those to our base layout: templates/base.html.twig.
I already have a couple of link tags up here in the stylesheets block. Add {{ encore_entry_link_tags() }} and pass it app, which is name of that entry.
This will render the built version of the app.css file plus any other CSS that our JavaScript imports. And actually, for performance reasons, Webpack may split that built app.css into multiple files: we'll see that in a few minutes. This function takes care of including all the link tags we need.
Do the same thing in the javascripts block: {{ encore_entry_script_tags() }} and pass app.
Like with the styles, this will render the built version of app.js plus any JavaScript it imports. And it may also be split into multiple files.

The New <script defer Attribute
Oh, and you're seeing a recent change in Symfony's TwigBundle recipe for base.html.twig. The javascripts block used to live down here at the bottom of the page. Now it lives up in the head in new projects. Normally... that would not be a good idea, because when your browser sees a script tag, it stops rendering the page until it can download and execute that JavaScript.
But thanks to a new feature in WebpackEncoreBundle - you can see it in config/packages/webpack_encore.yaml, here it is script_attributes - every script tag rendered by Encore will have a defer attribute. That basically means that the JavaScript isn't executed until after the page loads, very similar to having the scripts at the bottom of the page. But with the new setup, our browser does start downloading the files slightly earlier. And this plays nicer with Turbo - the topic of our next tutorial.
Iif you want to learn more about this change, check out a blog post I wrote on Symfony.com.
Next: let's use Encore to build our assets, see "code splitting" in action and import a third-party CSS package.