7.1. Actions: Listening to Events
I want to make our controller more realistic: instead of being able to click anywhere on the element to increment the count, let's add a button.
Easy enough! In the template, add <button class="btn btn-primary btn-sm"> and then the excellent call to action: "Click me".
If we check it in the browser... amazing! It works! Ok, I'm kidding: of course it works. But that's the cool thing about Stimulus: we get to do so much of our work in Twig where life is quick and easy.
Now: how could we attach a click listener to just this element? You might be thinking:
I know! Let's add a target to the button like we did for the span. Then we can find that button inside of connect() and add the event listener to it instead of this.element.
And yeah! That will totally do it! But it's way too much work.

Adding a Click Action
After targets, the second big feature of Stimulus is actions. Anytime you need to listen to an event, like click, submit, keyup or anything else, you can use an action instead.
Here's how it works: on the button element, I'll break this onto multiple lines for clarity. Now add data-action="click->counter#increment".
This is another special syntax: data-action="", then the name of the event, like click, submit or keyup, arrow, the name of the controller, a pound sign, and then the name of the method to call on our controller when this event happens. We'll create the increment method in a minute.
When I first used Stimulus, I did not love this syntax. It's... a bit weird. But it really does make life a lot nicer. And we'll simplify it a bit in a few minutes.
Over in the controller, add the method - increment() - then copy the logic from the click callback, delete, and paste it here.
And now we can remove the event listener entirely.
I may not love the data-action syntax in the template, but I do love the result. This is gorgeous.
Let's try it. Refresh the page. Now, if I click somewhere else in the element, nothing happens. If I click on the button, it increments! Woo!