8.1. Bootstrapping a "Color Selector" Form Element
Let's work on an even more interesting and complex example. Some products come in multiple colors. To select a color and quantity, I built a nice, boring Symfony form. If you're curious, you can see this form in src/Form/AddItemToCartForm.php. The color is an EntityType to a Color entity, which means it renders as a dropdown.
But... the design team isn't happy, and I don't blame them! They want something more interesting where users can actually see what the color looks when they they choose it.
So here's the goal: replace this select element with small "color square" boxes where we choose the color by clicking on the box. The strategy we're using here is called progressive enhancement. That's a philosophy where we get the page working first - with a nice boring, normal form - then make it better - "enhance it" - with JavaScript when we need to.
Open up the template for this page: it's down here at templates/product/_cart_add_controls.html.twig. If you're wondering, I extracted the form rendering code into this template "partial" because it's also used on the checkout sidebar.
Yup, this is a normal Symfony form that uses the normal Symfony form rendering functions.

Building the Color Boxes
Ok: to build the color boxes, should we add that HTML in Twig or in a Stimulus controller? The answer is almost always: in Twig. Why? Because Stimulus is all about rendering HTML on the server and adding behavior in JavaScript.
Since not all products come in multiple colors, this if statement checks to see if our form even has a color field. For forms that do have this field, each product has a different list of possible colors.
If you want to follow the entity structure, our form is bound to a CartItem object. We can go from CartItem to Product... and then once we have the Product, it has a colors property that holds a collection of Color objects that this product is available in. This is what we need to loop over to create the color boxes.
To do this, we'll use a little form trick: {% for color in addToCartForm.vars.data - that will give us the CartItem object - then .product.colors. And an endfor, then inside, we can say {{ color.hexColor }}: that's one of the properties on the Color object.
Let's try that. Move over and... ah! It works, but it's ugly.