8.3. Bootstrapping a "Color Selector" Form Element
Adding the Stimulus Controller
Now we need to make these little suckers functional. And that means we need a Stimulus controller.
Up in the assets/controllers/ directory, add a new file. Let's call it, how about, color-square_controller.js.
Notice the naming convention: the only thing that really matters is that we end in _controller.js. For the rest of the name, it could be color-square or color_square - it doesn't really matter, as long as it's all lowercase because the name is used in HTML attributes. Because I'm using dashes, the controller's name will be color-square.
Inside the file, we always start the same way: import { Controller } from 'stimulus' and export default class extends Controller.
I usually like to add a connect() method to make sure I've got everything hooked up correctly. Let's console.log(this.element.innerHTML)..
Now, go activate this in the template. But let's think: we need the controller to go around the three color boxes. But it also needs to go around the select element itself so that we can set its value when the user clicks a color box. The select element is rendered by form_widget().
So let's add a new <div data-controller="color-square">, put that around everything and indent.
Sweet! Let's take this puppy for a walk. Move over, refresh, and open up the console. Yes! Our controller is connected!