8.4. Bootstrapping a "Color Selector" Form Element
Adding the Action
Of course the goal of our Stimulus controller is not just to log something: it's to do something when we click each color square. So what we need is an action. In the template, on the color square span, add data-action="". Remember, the syntax for an action is the name of the controller - color-square - a #, then name of the method that should be called on our controller when this action happens. How about selectColor.
Now, over in the controller, replace the connect() method with selectColor() and, just like normal JavaScript, this will be passed an event object. Let's console.log() that event to see what it looks like.
Move over, refresh and... click! Uh... nothing happens? The action is not working!
This is one of the trickier things about Stimulus: often, if you have a slight mistake - like you misspell your controller name - you won't get an error... it just won't work. There are reasons for why.... but it can be tricky. Watch your spelling closely.
But in this case, the mistake is something else. I added the action to a span element. Things like a tags, buttons, forms and form elements all have a default action like click or submit. A span... does not... which makes sense: a span doesn't do anything in normal HTML.