8.6. Bootstrapping a "Color Selector" Form Element
Stay Semantic
But... we're making our life harder than it needs to be! Why not just make these squares buttons instead? That'll simplify the action syntax and... it's just more correct: these really are buttons that the user will click.
Change the span to a button. And then add type="button".
That will make sure that the button doesn't cause the form around it to submit when we click. And then, we do not need the click anymore: that is the default action for a button.
By the way, since our button doesn't have any text in it, to make this more accessible, we should add an aria-label attribute for screen readers, like aria-label="choose the color red".
Anyways, let's try this! Refresh, click and... woohoo!
Now that things are set up, let's actually... yea know, do something on click. First, we need to add a border to whichever square the user clicks so that it looks "selected". And second, we need to set the value on this select element so that the selected color is submitted with the form.
Let's work on the first part next and learn about the "current target" property on events.