9.1. Actions & currentTarget
When we click a square, we need to add a border around the square to show that it's currently selected. I've already created a CSS class for this... I'll hack it into the page so you can see it. It's called selected. It even comes with a nice little CSS transition! Ooooo.
Over in the controller, in the selectColor() method, how can we figure out which of the three color squares was just clicked? The answer is always: event.currentTarget.
Try this: event.currentTarget.classList.add('selected').
Before we chat about this, let's make sure it works. Refresh, click and... beautiful! We can currently select multiple colors... which isn't ideal, but we'll fix that soon.