9.2. Actions & currentTarget
currentTarget versus target
There are two important things about this line. First, when you listen to an event - or "action" in stimulus - the event object always has two similar properties: event.target and event.currentTarget. Sometimes these are the same element... and sometimes they're not.
Let me show you an example with some dummy code in the template. Imagine you have a button with an action on it - I'll reuse our existing data-action. Inside the button we have some text... but some of that text is inside another element. A more realistic example might be that you have an image or FontAwesome icon inside.
In the controller, I'll temporarily comment-out our code and instead console.log() event.target and also event.currentTarget so we can see the difference.
Go refresh the page. There's our stunning button! Open up the console.
First, click the text that's directly in the button. Nice! both event.target and event.currentTarget are the same thing: the button element.
Now click the the span that's inside the button. Woh! This time they're different! The target is the span while currentTarget is still the button!
This is not a Stimulus thing: this is just how DOM events work. event.target will always be the actual element that received the action. The second time we clicked, we were actually clicking the span.
But event.currentTarget will always be the element that we added the listener or action to.
So that's a long way of saying that event.currentTarget is your friend, because it will return the element that we've attached our action to. So we always know what it's going to be. event.target could be that element... or it could be a child element.