9.3. Actions & currentTarget
Let me remove that weird extra button... and then put the code back in our controller.

Element.classList
The other interesting thing on the line in our controller is classList. This is a property on the native Element object and... as you can see, it's just an easy way to add or remove class. No jQuery or other fancy tools needed.

Only Allowing One Selector
So... our color selector works great so far. Except for the problem that we can select multiple colors. We need to make sure that only one color has the selected class at a time.
Let's think about how to solve this. One option would be to look for an element with the selected class inside this.element. And if we find one, remove the class.
Another option is to use a target. We could make each color square a target, then, on click, loop over all of them and remove the selected class before re-adding it to the one that was just clicked.
Let's do that. First, define the target with targets = []. Let's call the new target, how about, colorSquare. I did just make a mistake: see if you can spot it.
Oh, and notice the naming of the target: it's lower camel case. I'm not using color-square because the name of the target becomes a property.
Down in the method, let's console.log(this.colorSquareTargets).
I put an "s" on the end on purpose: this will return an array of all matching targets.
Finally, in the template, let's add the target to the button. Remember: the syntax for that is data- the name of the controller - so color-square - the word target equals, then the name of the target: colorSquare.
Yes, you do need to write a few targets before you remember this syntax by heart. But you'll get it.
Let's try this. Refresh, click and... oh! Undefined?
Hopefully... you saw my mistake. Back in the controller, make this static targets. I made that mistake because... in the real world... I've made that mistake more than a few times before. This must be static and... if you forget, there's no huge error: it just won't add the magic target properties.
Try it now. Refresh, click and... yes! We see the 3 button elements.