10.1. Element.dataset
To make our widget fully work, when we click a color, we need to change the real select element's value. That way, when we submit, the form will POST the correct color. Inspect element near the select and look at its options: each value is the id of that color in the database. So somehow, when we click a color square, we need to know the id of that color so we can set it as the select element's value.

Adding our Own Custom data- Attributes
Fortunately, there's a native way to add extra information to DOM elements: data attributes!
Over in the template, find the button and add a new data attribute: data-color-id equals {{ color.id }}.
This is the first time that we've used a data attribute that has nothing to do with Stimulus. We're just inventing this for our own purposes. The only rules about data attribute names is that they must, of course, start with data- and they must be all lowercase. You also usually see dashes between words like color-id.
Over in our controller, we could read the attribute manually... but we don't have to! JavaScript has a built-in way to read data- attributes: it's the dataset property.
At the bottom of selectColor(), console.log(event.currentTarget) - to get the button - then .dataset.colorId.
Notice that the color-id from the HTML becomes colorId inside this dataset property. This is... once again, not a Stimulus thing. This is just how data attributes work.
Let's test it out. Open the console and... when I click, yes! We see ids!