10.2. Element.dataset
Adding the select Target
Now that we've got that, the next step is to find the select element. And, whenever we need to find something, it means that we need a target.
Over in the controller, add a second target called, how about, select.
Then, in the HTML, add that target. Oh... but this is trickier because the form_widget() function is rendering the select element for us. No problem: we can pass a custom attribute. Add a second argument to form widget, pass an associative array, give this an attr key set to another associative array with data-color-square-target set to select.
Back over in the controller, assuming I haven't messed anything up, we should now be able to reference the select with this.selectTarget. Set its value with .value = and then event.currentTarget.dataset.colorId.
Ok! Let's try this thing! Refresh click and... awesome! As we click the colors, the select updates. This is fun!
At this point... we're done! This will work! To celebrate, let's hide the select element. We could do that in Twig or in Stimulus - it's up to you. If you wanted to make your site work with and without JavaScript for some reason, you could hide the select element and show the color boxes at the same time in Stimulus.
Anyways, in the controller, add a connect() method. Then hide it with this.selectTarget.classList.add('d-none'), which will add a display: none since we're using Bootstrap.
Go refresh. Oh, that is lovely. Let's add a red sofa to the cart. When we submit... I think it worked! Go check out the shopping cart. It did!
Now that this is working, go check out our Stimulus controller. Yup, this whole feature required about 15 lines of JavaScript! That's thanks to the fact that all of our markup gets to live in Twig. Then, our JavaScript can stay lean and mean.
Over in the browser, click to go back to the sofa page. Next: I want to allow the user to click a color again to unselect that color. That won't be too hard, but to make it even easier, let's take advantage of the fact that we can store state on our controller.